+++ ESFramework/Foundation/NSString+ESExtension.h
@@ -13,21 +13,6 @@ NS_ASSUME_NONNULL_BEGIN
 @interface NSString (ESExtension)
 
 /**
- * Returns a NSNumber object created by parsing the string.
- */
-- (nullable NSNumber *)numberValue;
-
-@property (readonly) char charValue;
-@property (readonly) unsigned char unsignedCharValue;
-@property (readonly) short shortValue;
-@property (readonly) unsigned short unsignedShortValue;
-@property (readonly) unsigned int unsignedIntValue;
-@property (readonly) long longValue;
-@property (readonly) unsigned long unsignedLongValue;
-@property (readonly) unsigned long long unsignedLongLongValue;
-@property (readonly) NSUInteger unsignedIntegerValue;
-
-/**
  * Returns a NSData object containing a representation of the string encoded using UTF-8 encoding.
  */
 - (nullable NSData *)dataValue;