+++ ESFramework/Foundation/NSString+ESExtension.m
@@ -8,105 +8,11 @@
 
 #import "NSString+ESExtension.h"
 #import "ESHelpers.h"
-#import "NSNumber+ESExtension.h"
 #import "NSCharacterSet+ESExtension.h"
 #import "NSURLComponents+ESExtension.h"
 #import "NSData+ESExtension.h"
 
 @implementation NSString (ESExtension)
-
-+ (void)load
-{
-    static dispatch_once_t onceToken;
-    dispatch_once(&onceToken, ^{
-        ESSwizzleInstanceMethod(self, @selector(doubleValue), @selector(es_doubleValue));
-        ESSwizzleInstanceMethod(self, @selector(floatValue), @selector(es_floatValue));
-        ESSwizzleInstanceMethod(self, @selector(intValue), @selector(es_intValue));
-        ESSwizzleInstanceMethod(self, @selector(integerValue), @selector(es_integerValue));
-        ESSwizzleInstanceMethod(self, @selector(longLongValue), @selector(es_longLongValue));
-        ESSwizzleInstanceMethod(self, @selector(boolValue), @selector(es_boolValue));
-    });
-}
-
-- (double)es_doubleValue
-{
-    return self.numberValue.doubleValue;
-}
-
-- (float)es_floatValue
-{
-    return self.numberValue.floatValue;
-}
-
-- (int)es_intValue
-{
-    return self.numberValue.intValue;
-}
-
-- (NSInteger)es_integerValue
-{
-    return self.numberValue.integerValue;
-}
-
-- (long long)es_longLongValue
-{
-    return self.numberValue.longLongValue;
-}
-
-- (BOOL)es_boolValue
-{
-    return self.numberValue.boolValue;
-}
-
-- (NSNumber *)numberValue
-{
-    return [NSNumber numberWithString:self];
-}
-
-- (char)charValue
-{
-    return self.numberValue.charValue;
-}
-
-- (unsigned char)unsignedCharValue
-{
-    return self.numberValue.unsignedCharValue;
-}
-
-- (short)shortValue
-{
-    return self.numberValue.shortValue;
-}
-
-- (unsigned short)unsignedShortValue
-{
-    return self.numberValue.unsignedShortValue;
-}
-
-- (unsigned int)unsignedIntValue
-{
-    return self.numberValue.unsignedIntValue;
-}
-
-- (long)longValue
-{
-    return self.numberValue.longValue;
-}
-
-- (unsigned long)unsignedLongValue
-{
-    return self.numberValue.unsignedLongValue;
-}
-
-- (unsigned long long)unsignedLongLongValue
-{
-    return self.numberValue.unsignedLongLongValue;
-}
-
-- (NSUInteger)unsignedIntegerValue
-{
-    return self.numberValue.unsignedIntegerValue;
-}
 
 - (nullable NSData *)dataValue
 {